backend/requirements.txt
python-dotenv
PyMuPDF
Flask
flask-cors
Flask-Session
faiss-cpu
langchain==0.3.25
langchain-community
huggingface-hub==0.31.1
langchain-huggingface==0.2.0